requirements.txt
# OpenPrinterAgent - Production Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
Flask==3.1.2
flask-cors==4.0.0

# Printer Communication
python-escpos==3.1
pyserial==3.5
pyusb==1.3.1

# Image Processing & Generation
pillow==12.0.0
python-barcode==0.16.1
qrcode==8.2

# Desktop GUI
customtkinter==5.2.2

# Build / Packaging
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11

# Utilities
PyYAML==6.0.3
requests==2.32.5
python-dotenv==1.0.0
